test(pie): fix blank visual snapshots for data label percentages
Wait for chart render and move custom formatter into a dedicated story
component so Playwright CT captures the pie with labels.
